+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### v2.20.0 (2023-05-16)
+* * *
+
+#### New Feature:
+* Added setIdempotencyKey("UUID") utility to pass **Idempotency key** along with request headers to allow a safe retry of POST requests.
+* Added isIdempotencyReplayed() utility to differentiate between original and replayed requests.
+* Added getResponseHeaders() utility to to fetch the response headers.
+
 ### v2.19.0 (2023-04-28)
 * * *